Add select2 jquery & validate delete

+++ app/Http/Controllers/CategoriesController.php
@@ -105,6 +105,12 @@ class CategoriesController extends Controller
      */
     public function destroy(Categories $category)
     {
+        if($category->posts()->count() > 0){
+
+            session()->flash('error','Category cannot be deleted , some post still exist');
+
+            return redirect()->back();
+        }
         $category->delete();
 
         session()->flash('success','Delete successfully');
