+++ app/Http/Controllers/TagsController.php
@@ -96,6 +96,13 @@ class TagsController extends Controller
      */
     public function destroy(Tag $tag)
     {
+
+        if($tag->posts()->count() > 0){
+
+            session()->flash('error','Tag cannot be deleted , some post still exist');
+
+            return redirect()->back();
+        }
         $tag->delete();
 
         session()->flash('success','Tag delete successfully');
